python/webapp/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.3
